+++ test/runner.js
@@ -1,0 +1,4 @@
+import chai from 'chai'
+chai.should()
+
+require('./resolve')